requirements.txt
psycopg2
groq
uvicorn
fastapi
python-dotenv
streamlit